pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/cdk_factory"]

[tool.pytest.ini_options]
pythonpath = ["src", "samples"]
testpaths = ["tests/unit"]
addopts = "-m 'not integration'"
markers = [
    "integration: marks tests as integration (deselect with '-m \"not integration\"')"
]

[project]

name = "cdk_factory"
version = "0.0.10"
authors = [
  { name="Eric Wilson", email="eric.wilson@geekcafe.com" }
]
description = "CDK Factory. A QuickStarter and best practices setup for CDK projects"
readme = "README.md"
license = {text = "MIT License"}
requires-python = ">=3.10"
classifiers = [
    "Programming Language :: Python :: 3",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
]

dependencies = [
  "aws-lambda-powertools>=2.38.1",
  "aws-cdk-lib",
  "constructs",
  "boto3_assist>=0.10.0",
  "PyYAML"
]